Split testRoleDownloadsAreLimited() into its own class
